pyproject.toml
[tool.poetry]
name = "django_scoped_permissions"
version = "0.1.0"
description = "Test scoped permissions"
authors = ["Magnus Buvarp <magnus.buvarp@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.3"
graphene-django = "^2.8.0"
graphene-django-cud = "^0.3.0"
pyjwt = "^1.7.1"
factory_boy = "^2.12.0"

[tool.poetry.dev-dependencies]
addict = "^2.2.1"
pytest = "^5.3.5"
pytest-django = "^3.8.0"
pytest-watch = "^4.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
